Adds filtering functionality to character list
Implements filtering for character lists based on character type and species.

Introduces a filter dropdown component for desktop and a separate filter page for mobile, enhancing user experience.

Updates the search bar to include a filter icon, enabling users to easily access and apply filters.

Updates MainLayout to add a background color to the list container.